Enhanced User Management
git-svn-id: http://svn.uscm.org/si/trunk@13 68e5226f-b13b-0410-b3eb-f6601b5927d1

+++ app/controllers/users_controller.rb
@@ -14,26 +14,43 @@ class UsersController < ApplicationController
   end
   
   def edit
+   @temp_user = SiUser.find(params[:id])
+   
     respond_to do |format|
       format.html
     end
   end
   
   def create
-    @person = Person.find(:first, :conditions => ['personID = ?', params[:person_id]], :include => :user)
-    old_user = SiUser.find_by_ssm_id(@person.user) 
-    old_user.destroy if old_user # delete the old user so we can create the new one.
-    type = params[:user][:role]
-    @new_user = type.constantize.create!(:ssm_id => @person.user.id,
-                                         :created_at => Time.now,
-                                         :created_by_id => user.id)
-
-    redirect_to users_url
+    @temp_user = SiUser.new
+    if params[:person_id].blank?
+      @temp_user.errors.add_to_base "You must first select a person before creating a new user."
+      render :action => :new
+    else
+      @person = Person.find(:first, :conditions => ['personID = ?', params[:person_id]], :include => :user)
+      old_user = SiUser.find_by_ssm_id(@person.user) 
+      old_user.destroy if old_user # delete the old user so we can create the new one.
+      type = params[:temp_user][:role]
+      @new_user = type.constantize.create!(:ssm_id => @person.user.id,
+                                           :created_at => Time.now,
+                                           :created_by_id => user.id)
+  
+      redirect_to users_url
+    end
   end
   
   def update
     respond_to do |format|
       format.html
+    end
+  end
+  
+  def destroy
+    @temp_user = SiUser.find(params[:id])
+    @temp_user.destroy
+    
+    respond_to do |format|
+      format.html { redirect_to users_url }
     end
   end
   
@@ -44,7 +61,7 @@ class UsersController < ApplicationController
   end
   
   protected
-    def person_search(name, options = {})
+  def person_search(name, options = {})
     @people = Array.new
     if name and !name.empty?
       names = name.strip.split(' ')
@@ -60,5 +77,6 @@ class UsersController < ApplicationController
       @conditions[0] += " AND accountNo <> '' AND accountNo is NOT NULL " if options[:staff_only]
       @people = Person.find(:all, :order => "lastName, firstName", :conditions => @conditions)
     end
+    return @people
   end
 end
